@@ -1,21 +1,20 @@
 import 'dart:io';
-
+import 'dart:math';
 import 'package:flutter/material.dart';
 
 class Message with ChangeNotifier {
-  final String id;
-  String _text;
-  DateTime _dateTime;
+  final int id = Random().nextInt(256);
+  String _text = "";
+  DateTime _dateTime = DateTime.now();
   final AnimationController animController;
-  final File imageFile;
+  final File? imageFile;
 
   Message(
-      {String text,
-      DateTime dateTime,
-      this.id,
-      this.animController,
+      {String? text,
+      DateTime? dateTime,
+      required this.animController,
       this.imageFile}) {
-    this._text = text;
+    this._text = text ?? "";
     this._dateTime = dateTime ?? DateTime.now();
   }
 
