@@ -1,13 +1,12 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class CustomIconButton extends IconButton {
   CustomIconButton({
-    Key key,
-    @required Widget icon,
-    @required VoidCallback onPressed,
-    VisualDensity visualDensity,
-    String tooltip,
+    Key? key,
+    required Widget icon,
+    required VoidCallback onPressed,
+    VisualDensity? visualDensity,
+    String? tooltip,
   }) : super(
             icon: icon,
             onPressed: onPressed,
