@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:image_picker/image_picker.dart';
 
 Future<File> getImage() async {
-  File _image;
+  File? _image;
   final picker = ImagePicker();
   final pickedFile = await picker.getImage(source: ImageSource.gallery);
 
@@ -12,5 +12,5 @@ Future<File> getImage() async {
   } else {
     print('@@@@@@@@@@ Not image selected.');
   }
-  return _image;
+  return _image!;
 }
